new direction for parsing. still confused.
