+++ src/utils/utils.h
@@ -1,0 +1,9 @@
+#ifndef UTILS_H
+# define UTILS_H
+
+# include "../minishell.h"
+
+// utils
+void	*malloc_perex(size_t bytes, char *msg);
+
+#endif
